ref: Flatten project structure further
Flatten project structure further to make imports more accessible.
