+++ decent_bench/__init__.py
@@ -1,0 +1,25 @@
+from decent_bench import (
+    agent,
+    benchmark,
+    benchmark_problem,
+    centralized_algorithms,
+    cost_functions,
+    datasets,
+    distributed_algorithms,
+    metrics,
+    network,
+    schemes,
+)
+
+__all__ = [
+    "agent",
+    "benchmark",
+    "benchmark_problem",
+    "centralized_algorithms",
+    "cost_functions",
+    "datasets",
+    "distributed_algorithms",
+    "metrics",
+    "network",
+    "schemes",
+]
