--- decent_bench/cost_functions/linear_regression_cost.py
@@ -1,109 +1,0 @@
-from numpy import float64
-from numpy.typing import NDArray
-
-from decent_bench.cost_functions.cost_function import CostFunction
-from decent_bench.cost_functions.quadratic_cost import QuadraticCost
-
-
-class LinearRegressionCost(CostFunction):
-    r"""
-    Linear regression cost function.
-
-    .. math:: f(\mathbf{x}) = \frac{1}{2} \| \mathbf{Ax} - \mathbf{b} \|^2
-
-    or in the general quadratic form
-
-    .. math::
-        f(\mathbf{x})
-        = \frac{1}{2} \mathbf{x}^T\mathbf{A}^T\mathbf{Ax}
-        - (\mathbf{A}^T \mathbf{b})^T \mathbf{x}
-        + \frac{1}{2} \mathbf{b}^T\mathbf{b}
-    """
-
-    def __init__(self, A: NDArray[float64], b: NDArray[float64]):  # noqa: N803
-        if A.shape[0] != b.shape[0]:
-            raise ValueError(f"Dimension mismatch: A has {A.shape[0]} rows but b has {b.shape[0]} elements")
-        self.inner = QuadraticCost(A.T.dot(A), -A.T.dot(b), 0.5 * b.dot(b))
-        self.A = A
-        self.b = b
-
-    @property
-    def domain_shape(self) -> tuple[int, ...]:  # noqa: D102
-        return self.inner.domain_shape
-
-    @property
-    def m_smooth(self) -> float:
-        r"""
-        The cost function's smoothness constant.
-
-        .. math::
-            \max_{i} \left| \lambda_i \right|
-
-        where :math:`\lambda_i` are the eigenvalues of :math:`\mathbf{A}^T \mathbf{A}`.
-
-        For the general definition, see
-        :attr:`CostFunction.m_smooth <decent_bench.cost_functions.cost_function.CostFunction.m_smooth>`.
-        """
-        return self.inner.m_smooth
-
-    @property
-    def m_cvx(self) -> float:
-        r"""
-        The cost function's convexity constant.
-
-        .. math::
-            \begin{array}{ll}
-                \min_i \lambda_i, & \text{if } \min_i \lambda_i > 0, \\
-                0, & \text{if } \min_i \lambda_i = 0, \\
-                \text{NaN}, & \text{if } \min_i \lambda_i < 0
-            \end{array}
-
-        where :math:`\lambda_i` are the eigenvalues of :math:`\mathbf{A}^T \mathbf{A}`.
-
-        For the general definition, see
-        :attr:`CostFunction.m_cvx <decent_bench.cost_functions.cost_function.CostFunction.m_cvx>`.
-        """
-        return self.inner.m_cvx
-
-    def evaluate(self, x: NDArray[float64]) -> float:
-        r"""
-        Evaluate function at x.
-
-        .. math:: \frac{1}{2} \| \mathbf{Ax} - \mathbf{b} \|^2
-        """
-        return self.inner.evaluate(x)
-
-    def gradient(self, x: NDArray[float64]) -> NDArray[float64]:
-        r"""
-        Gradient at x.
-
-        .. math:: \mathbf{A}^T\mathbf{Ax} - \mathbf{A}^T \mathbf{b}
-        """
-        return self.inner.gradient(x)
-
-    def hessian(self, x: NDArray[float64]) -> NDArray[float64]:
-        r"""
-        Hessian at x.
-
-        .. math:: \mathbf{A}^T\mathbf{A}
-        """
-        return self.inner.hessian(x)
-
-    def proximal(self, y: NDArray[float64], rho: float) -> NDArray[float64]:
-        r"""
-        Proximal at y.
-
-        .. math::
-            (\rho \mathbf{A}^T \mathbf{A} + \mathbf{I})^{-1} (\mathbf{y} + \rho \mathbf{A}^T\mathbf{b})
-
-        where :math:`\rho > 0` is the penalty.
-
-        This is a closed form solution, see
-        :meth:`CostFunction.proximal() <decent_bench.cost_functions.cost_function.CostFunction.proximal>`
-        for the general proximal definition.
-        """
-        return self.inner.proximal(y, rho)
-
-    def __add__(self, other: CostFunction) -> CostFunction:
-        """Add another cost function."""
-        return self.inner + other
