--- decent_bench/cost_functions/quadratic_cost.py
@@ -1,129 +1,0 @@
-import numpy as np
-from numpy import float64
-from numpy.typing import NDArray
-
-from decent_bench.cost_functions.cost_function import CostFunction
-from decent_bench.cost_functions.sum_cost import SumCost
-
-
-class QuadraticCost(CostFunction):
-    r"""
-    Quadratic cost function.
-
-    .. math:: f(\mathbf{x}) = \frac{1}{2} \mathbf{x}^T \mathbf{Ax} + \mathbf{b}^T \mathbf{x} + c
-    """
-
-    def __init__(self, A: NDArray[float64], b: NDArray[float64], c: float):  # noqa: N803
-        if A.ndim != 2:
-            raise ValueError("Matrix A must be 2D")
-        if A.shape[0] != A.shape[1]:
-            raise ValueError("Matrix A must be square")
-        if b.ndim != 1:
-            raise ValueError("Vector b must be 1D")
-        if A.shape[0] != b.shape[0]:
-            raise ValueError(f"Dimension mismatch: A has shape {A.shape} but b has length {b.shape[0]}")
-        self.A = A
-        self.A_sym = 0.5 * (A + A.T)
-        self.b = b
-        self.c = c
-
-    @property
-    def domain_shape(self) -> tuple[int, ...]:  # noqa: D102
-        return self.b.shape
-
-    @property
-    def m_smooth(self) -> float:
-        r"""
-        The cost function's smoothness constant.
-
-        .. math::
-            \max_{i} \left| \lambda_i \right|
-
-        where :math:`\lambda_i` are the eigenvalues of :math:`\frac{1}{2} (\mathbf{A}+\mathbf{A}^T)`.
-
-        For the general definition, see
-        :attr:`CostFunction.m_smooth <decent_bench.cost_functions.cost_function.CostFunction.m_smooth>`.
-        """
-        eigs = np.linalg.eigvalsh(self.A_sym)
-        return float(np.max(np.abs(eigs)))
-
-    @property
-    def m_cvx(self) -> float:
-        r"""
-        The cost function's convexity constant.
-
-        .. math::
-            \begin{array}{ll}
-                \min_i \lambda_i, & \text{if } \min_i \lambda_i > 0, \\
-                0, & \text{if } \min_i \lambda_i = 0, \\
-                \text{NaN}, & \text{if } \min_i \lambda_i < 0
-            \end{array}
-
-        where :math:`\lambda_i` are the eigenvalues of :math:`\frac{1}{2} (\mathbf{A}+\mathbf{A}^T)`.
-
-        For the general definition, see
-        :attr:`CostFunction.m_cvx <decent_bench.cost_functions.cost_function.CostFunction.m_cvx>`.
-        """
-        eigs = np.linalg.eigvalsh(self.A_sym)
-        l_min = float(np.min(eigs))
-        tol = 1e-12
-        if l_min > tol:
-            return l_min
-        if abs(l_min) <= tol:
-            return 0
-        return np.nan
-
-    def evaluate(self, x: NDArray[float64]) -> float:
-        r"""
-        Evaluate function at x.
-
-        .. math:: \frac{1}{2} \mathbf{x}^T \mathbf{Ax} + \mathbf{b}^T \mathbf{x} + c
-        """
-        return float(0.5 * x.dot(self.A.dot(x)) + self.b.dot(x) + self.c)
-
-    def gradient(self, x: NDArray[float64]) -> NDArray[float64]:
-        r"""
-        Gradient at x.
-
-        .. math:: \frac{1}{2} (\mathbf{A}+\mathbf{A}^T)\mathbf{x} + \mathbf{b}
-        """
-        return self.A_sym @ x + self.b
-
-    def hessian(self, x: NDArray[float64]) -> NDArray[float64]:  # noqa: ARG002
-        r"""
-        Hessian at x.
-
-        .. math:: \frac{1}{2} (\mathbf{A}+\mathbf{A}^T)
-        """
-        return self.A_sym
-
-    def proximal(self, y: NDArray[float64], rho: float) -> NDArray[float64]:
-        r"""
-        Proximal at y.
-
-        .. math::
-            (\frac{\rho}{2} (\mathbf{A} + \mathbf{A}^T) + \mathbf{I})^{-1} (\mathbf{y} - \rho \mathbf{b})
-
-        where :math:`\rho > 0` is the penalty.
-
-        This is a closed form solution, see
-        :meth:`CostFunction.proximal() <decent_bench.cost_functions.cost_function.CostFunction.proximal>`
-        for the general proximal definition.
-        """
-        lhs = rho * self.A_sym + np.eye(self.A.shape[1])
-        rhs = y - self.b * rho
-        return np.asarray(np.linalg.solve(lhs, rhs), dtype=float64)
-
-    def __add__(self, other: CostFunction) -> CostFunction:
-        """
-        Add another cost function.
-
-        Raises:
-            ValueError: if the domain shapes don't match
-
-        """
-        if self.domain_shape != other.domain_shape:
-            raise ValueError(f"Mismatching domain shapes: {self.domain_shape} vs {other.domain_shape}")
-        if isinstance(other, QuadraticCost):
-            return QuadraticCost(self.A + other.A, self.b + other.b, self.c + other.c)
-        return SumCost([self, other])
