+++ decent_bench/metrics/__init__.py
@@ -1,0 +1,3 @@
+from . import metric_utils, plot_metrics, table_metrics
+
+__all__ = ["metric_utils", "plot_metrics", "table_metrics"]
