+++ decent_bench/utils/progress_bar.py
@@ -5,7 +5,7 @@ from threading import Thread
 
 from rich.progress import BarColumn, Progress, TaskID, TaskProgressColumn, TextColumn, TimeRemainingColumn
 
-from decent_bench.algorithms.dst_algorithms import DstAlgorithm
+from decent_bench.distributed_algorithms import DstAlgorithm
 
 
 @dataclass(eq=False)
