--- docs/source/decent_bench.algorithms.rst
@@ -1,29 +1,0 @@
-decent\_bench.algorithms package
-================================
-
-Submodules
-----------
-
-decent\_bench.algorithms.cent\_algorithms module
-------------------------------------------------
-
-.. automodule:: decent_bench.algorithms.cent_algorithms
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.algorithms.dst\_algorithms module
------------------------------------------------
-
-.. automodule:: decent_bench.algorithms.dst_algorithms
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: decent_bench.algorithms
-   :members:
-   :show-inheritance:
-   :undoc-members:
