--- docs/source/decent_bench.datasets.rst
@@ -1,29 +1,0 @@
-decent\_bench.datasets package
-==============================
-
-Submodules
-----------
-
-decent\_bench.datasets.dataset module
--------------------------------------
-
-.. automodule:: decent_bench.datasets.dataset
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.datasets.synthetic\_classification\_data module
--------------------------------------------------------------
-
-.. automodule:: decent_bench.datasets.synthetic_classification_data
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: decent_bench.datasets
-   :members:
-   :show-inheritance:
-   :undoc-members:
