--- docs/source/decent_bench.metrics.plot_metrics.rst
@@ -1,45 +1,0 @@
-decent\_bench.metrics.plot\_metrics package
-===========================================
-
-Submodules
-----------
-
-decent\_bench.metrics.plot\_metrics.default\_plot\_metrics module
------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.plot_metrics.default_plot_metrics
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.plot\_metrics.plot module
------------------------------------------------
-
-.. automodule:: decent_bench.metrics.plot_metrics.plot
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.plot\_metrics.plot\_metrics\_constructs module
---------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.plot_metrics.plot_metrics_constructs
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.plot\_metrics.plot\_metrics\_data\_extractors module
---------------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.plot_metrics.plot_metrics_data_extractors
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: decent_bench.metrics.plot_metrics
-   :members:
-   :show-inheritance:
-   :undoc-members:
