+++ docs/source/decent_bench.metrics.rst
@@ -1,15 +1,6 @@
 decent\_bench.metrics package
 =============================
 
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   decent_bench.metrics.plot_metrics
-   decent_bench.metrics.table_metrics
-
 Submodules
 ----------
 
@@ -17,6 +8,22 @@ decent\_bench.metrics.metric\_utils module
 ------------------------------------------
 
 .. automodule:: decent_bench.metrics.metric_utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.metrics.plot\_metrics module
+------------------------------------------
+
+.. automodule:: decent_bench.metrics.plot_metrics
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.metrics.table\_metrics module
+-------------------------------------------
+
+.. automodule:: decent_bench.metrics.table_metrics
    :members:
    :show-inheritance:
    :undoc-members:
