--- docs/source/decent_bench.metrics.table_metrics.rst
@@ -1,45 +1,0 @@
-decent\_bench.metrics.table\_metrics package
-============================================
-
-Submodules
-----------
-
-decent\_bench.metrics.table\_metrics.default\_table\_metrics module
--------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.table_metrics.default_table_metrics
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.table\_metrics.table\_metrics\_constructs module
-----------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.table_metrics.table_metrics_constructs
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.table\_metrics.table\_metrics\_data\_extractors module
-----------------------------------------------------------------------------
-
-.. automodule:: decent_bench.metrics.table_metrics.table_metrics_data_extractors
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.metrics.table\_metrics.tabulate module
-----------------------------------------------------
-
-.. automodule:: decent_bench.metrics.table_metrics.tabulate
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: decent_bench.metrics.table_metrics
-   :members:
-   :show-inheritance:
-   :undoc-members:
