--- docs/source/decent_bench.schemes.rst
@@ -1,45 +1,0 @@
-decent\_bench.schemes package
-=============================
-
-Submodules
-----------
-
-decent\_bench.schemes.agent\_activation\_schemes module
--------------------------------------------------------
-
-.. automodule:: decent_bench.schemes.agent_activation_schemes
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.schemes.compression\_schemes module
--------------------------------------------------
-
-.. automodule:: decent_bench.schemes.compression_schemes
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.schemes.drop\_schemes module
-------------------------------------------
-
-.. automodule:: decent_bench.schemes.drop_schemes
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-decent\_bench.schemes.noise\_schemes module
--------------------------------------------
-
-.. automodule:: decent_bench.schemes.noise_schemes
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: decent_bench.schemes
-   :members:
-   :show-inheritance:
-   :undoc-members:
