+++ docs/source/decent_bench.utils.rst
@@ -1,0 +1,29 @@
+decent\_bench.utils package
+===========================
+
+Submodules
+----------
+
+decent\_bench.utils.logger module
+---------------------------------
+
+.. automodule:: decent_bench.utils.logger
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.utils.progress\_bar module
+----------------------------------------
+
+.. automodule:: decent_bench.utils.progress_bar
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: decent_bench.utils
+   :members:
+   :show-inheritance:
+   :undoc-members:
