+++ test/test_import.py
@@ -1,2 +1,3 @@
 def test_import():
     import decent_bench
+    import decent_bench.benchmark
